Added a WIP function to clear tags
Basically, a few specific tags (`ITUNESADVISORY`, `TRACKTOTAL` and `DISCTOTAL`) tend to append to old values.
This has so far been observed for MP3 only - other formats have so far been untested

What if we remove the tag before we write to it? That's exactly this commit's purpose.

Additionally, added another sample track :3