Start work on newer tag duplication fix
This problem so far seems exclusive to ID3 tags. When I opened an MP3 file with `mutagen.mp3.MP3`, I found `TXXX:itunesadvisory` and `TXXX:ITUNESADVISORY` to be different tags (basically, I learnt that ID3 tags are case-sensitive) - `mutagen.mp3.EasyMP3` (a wrapper for `mutagen.mp3.MP3`) sees them as the same and for some reason finds only one of them. This will be a fun amount of work :3